requirements.txt
pip~=23.2.1
wheel~=0.38.4
Jinja2~=3.1.2
setuptools~=65.5.1
Flask~=2.3.2
Werkzeug~=2.3.6
flask
flask-smorest
python-dotenv
Pillow~=10.0.0
packaging~=23.1
requests~=2.31.0